:sparkles: Feat: 게시판 플로팅 버튼 추가

+++ project/Myo_jib_sa/app/src/main/java/com/example/myo_jib_sa/community/api/communityHome/HomeResponse.kt
@@ -8,7 +8,7 @@ data class HomeResponse (
 data class HomeResult(
     val mainMission:List<MainMission>,
     val popularArticle:List<PopularArticle>,
-    val userHostCategory:List<Int>
+    val userHostCategory:List<Long>
 )
 data class MainMission(
     val mainMissionId:Long,
